split midpoint/dual midpoint
